⚡️ :: 이름 변경

+++ src/main/java/com/gsp/bridge/domain/notification/domain/NotificationListId.java
@@ -13,6 +13,6 @@ public class NotificationListId implements Serializable {
 
     private Integer notificationEntity;
 
-    private Long user;
+    private Integer user;
 
 }
